Add more type info

+++ src/Composer/Repository/CompositeRepository.php
@@ -23,13 +23,13 @@ class CompositeRepository implements RepositoryInterface
 {
     /**
      * List of repositories
-     * @var array
+     * @var RepositoryInterface[]
      */
     private $repositories;
 
     /**
      * Constructor
-     * @param array $repositories
+     * @param RepositoryInterface[] $repositories
      */
     public function __construct(array $repositories)
     {
@@ -169,8 +169,9 @@ class CompositeRepository implements RepositoryInterface
     public function removePackage(PackageInterface $package)
     {
         foreach ($this->repositories as $repository) {
-            /* @var $repository RepositoryInterface */
-            $repository->removePackage($package);
+            if ($repository instanceof WritableRepositoryInterface) {
+                $repository->removePackage($package);
+            }
         }
     }
